requirements.txt
kornia
numpy
omegaconf
onnxruntime-gpu==1.18.1
torch
tqdm
